Moved 'ClassGenerator' to 'Generator' namespace and added unit test.
